requirements.txt
django==6.0
djangorestframework==3.16.1
celery==5.6.0
redis==5.2.0
boto3
django-storages
pypdf==6.4.1
pytesseract==0.3.13
pdf2image
django-allauth==65.13.1
python-dotenv
gunicorn
psycopg2-binary
dj-database-url
requests
requests-oauthlib
PyJWT
cryptography
google-generativeai
numpy


